interface DatabaseConnector
Create interface, rename abstract class to AbstractDatabaseConnector.
